RoundAndSwap.jl

RoundAndSwap.jl is a library for implementing the Round and Swap algorithm to try and find an intial solution.

Getting started

Create your JuMP model, and optimize

# Define how many variables you want fixed
num_to_fix = 2

using JuMP, HiGHS
model = Model(HiGHS.Optimizer)
@variable(model, 0 ≤ a ≤ 1)
@variable(model, 0 ≤ b ≤ 1)
@variable(model, 0 ≤ c ≤ 1)
@variable(model, 0 ≤ d ≤ 1)
@constraint(model, a + b + c + d == num_to_fix)
@objective(model, Max, (a + b) + (2 * (b + c)) + (3 * (c - d)) + (4 * (d + a)))

# To demonstate functionality we will fix b and d
fix(b, 0.8; force=true)
fix(d, 0.8; force=true)

optimize!(model)

Identify which variable we are considering making 1

consider_swapping = [a,b,c,d]

Round variables closest to 1, this will be b and d as we fixed them at 0.8.

using RoundAndSwap
round!(consider_swapping, num_to_fix)

Begin swapping

best_swap, swapper = swap(model, consider_swapping)

Review your best swap

julia> best_swap
1-element Vector{Swap}:
 Swap 
    existing:           b
    new:                a
    obj_value:          10.0          
    success:            true            
    all_fixed:          [:a, :c]          
    termination_status: OPTIMAL 
    solve_time:         0.0004438320000872409
    swap_number:        7

As you can see in this case we found the globally optimal solution of 10. Note: Round and Swap does not provide guarantees of global optimility.

caveats

Currently can only set variables to 0 or 1.

Performance

In testing I've done 536 swaps with round_and_swap taking 174.1 seconds with 172.8 seconds of that being in the optimizer, giving RoundAndSwap.jl an overhead of 0.75%.

Nomenclature

Index

RoundAndSwap.Swap

RoundAndSwap.Swapper

Base.:==

Base.getindex

RoundAndSwap._completed_swaps

RoundAndSwap.best_objective

RoundAndSwap.best_swap

RoundAndSwap.create_swaps!

RoundAndSwap.create_swaps!

RoundAndSwap.evalute_sweep

RoundAndSwap.fix!

RoundAndSwap.fixed_variables

RoundAndSwap.fixed_variables

RoundAndSwap.flatten

RoundAndSwap.get_var

RoundAndSwap.initial_swaps

RoundAndSwap.make_models

RoundAndSwap.num_swaps

RoundAndSwap.previously_tried

RoundAndSwap.reduce_to_consider_number

RoundAndSwap.reduce_to_consider_percentile

RoundAndSwap.solve!

RoundAndSwap.status_codes

RoundAndSwap.successful

RoundAndSwap.successful

RoundAndSwap.successful_swaps

RoundAndSwap.swap

RoundAndSwap.swap

RoundAndSwap.swap

RoundAndSwap.total_optimisation_time

RoundAndSwap.try_swapping!

RoundAndSwap.unfix!

RoundAndSwap.unfix!

RoundAndSwap.unsuccessful_swaps

Functions

RoundAndSwap.Swap — Type

Swap

An objecct to keep track of a swap

Arguments:

existing::Union{Symbol, Nothing}: The existing variable

new::Union{Symbol, Nothing}: The variable to replace the existing

obj_value::Real: The objective value for this swap

success::Union{Bool, Nothing}: Whether the swap was successful

all_fixed::Union{Array{Symbol}, Nothing}: All the variable in swapper.to_consider which were fixed

termination_status::Union{String, TerminationStatusCode, Nothing}: TerminationStatusCode

solve_time::Union{Real, Nothing}: Time to solve

Swap(existing, new): A constructor for a Swap object

RoundAndSwap.Swapper — Type

Swapper

An object to keep track of all the swaps

Arguments:

to_swap::Array{Swap}: A list of swaps yet to be done

consider_swapping::Array{Symbol}: The variables to consider swapping

completed_swaps::Union{Array{Array{Swap}}, Nothing}: The completed swaps

sense::OptimizationSense: The optimisation sense, MAX or MIN

max_swaps::Real: The maximum number of swaps to complete

number_of_swaps::Int: The number of swaps completed

Swapper(to_swap, to_swap_with, model; max_swaps): A constructor

Base.:== — Method

Base.:(==)(a::Swap, b::Swap)

Check whther two swaps have the same existing and new values

Base.getindex — Method

Base.getindex(m::JuMP.AbstractModel, names::Array{Symbol})

Allow the getting of multiple variables

RoundAndSwap._completed_swaps — Method

_completed_swaps(swapper::Swapper)

Get a list of swaps which actually ran

RoundAndSwap.best_objective — Method

best_objective(swapper::Swapper)

Get the best objective value in swapper.completed_swaps

RoundAndSwap.best_swap — Method

best_swap(swapper::Swapper)

Return the best swap in the swapper

RoundAndSwap.create_swaps! — Method

create_swaps!(swapper::Swapper, to_swap::Array{Symbol})

Given the previously completed swaps, create a list of new swaps

RoundAndSwap.create_swaps! — Method

create_swaps!(swapper::Swapper, to_swap::Symbol)

Given the previously completed swaps, create a list of new swaps

RoundAndSwap.evalute_sweep — Method

evalute_sweep(swapper::Swapper)

After complete a sweep, find which swaps improved the existing best objective and use this to create new swaps

RoundAndSwap.fix! — Function

fix!(models::Array{Model}, to_fix::Array{Symbol}, value = 1)

For each model, fix all variables in to_fix to 1

Arguments:

models: Models to fix values

to_fix: Variables to fix

value: Value to fix to, by default 1

RoundAndSwap.fixed_variables — Method

fixed_variables(model::Model, consider_swapping::Array{Symbol})

Find which variables in consider swapping are fixed

RoundAndSwap.fixed_variables — Method

fixed_variables(model::Model, swapper::Swapper)

Find which variables in consider swapping are fixed

RoundAndSwap.flatten — Method

flatten(to_flatten::Array{Array{Swap}})

Flatten nested arrays of swaps

RoundAndSwap.get_var — Method

get_var(m::JuMP.AbstractModel, name::Symbol)

Get a variable, if the symbol fails try getting it from an array of variables

RoundAndSwap.initial_swaps — Method

initial_swaps(fixed_variables::Array{Symbol}, consider_swapping::Array{Symbol})

Given the initial state, create a list of initial swaps For all currently fixed, consider_swapping with all variables to consider wFixed = a consider = a,b,c,d Swaps([ a -> b, a -> c, a -> d ])

RoundAndSwap.make_models — Function

make_models(model::Model, optimizer::Union{Nothing, DataType} = nothing)

Given a single model, make enough models to have one for each thread

Arguments:

optimizer: Optimizer to use, if nothing, will use the same as the one in the provided model, proving it is one of [Gurobi, Ipopt, HiGHS]

RoundAndSwap.num_swaps — Method

num_swaps(swapper::Swapper)

Get how many swaps have been completed

RoundAndSwap.previously_tried — Method

previously_tried(swapper::Swapper)

Get a list of all previously stried fixed variables in swapper.consider_swapping

RoundAndSwap.reduce_to_consider_number — Method

reduce_to_consider(to_consider::Array{VariableRef}; num_desired_to_consider::Int)

Only consider the numdesiredtoconsider largest values in to consider. Note that toconsider shouldn't be rounded yet.

RoundAndSwap.reduce_to_consider_percentile — Method

reduce_to_consider(to_consider::Array{VariableRef}; percentile::Int)

Only consider the percentile largest values in to consider. Note that to_consider shouldn't be rounded yet.

RoundAndSwap.solve! — Method

solve!(model, swapper, swap)

Solve this swap

Arguments:

model: A model object

swapper: The swapper being used

swap: Which swap to solve

RoundAndSwap.status_codes — Method

status_codes(swapper::Swapper)

Get all the status codes which have occured for this Swapper

RoundAndSwap.successful — Method

successful(model::Model)

Determine if a model was succesful

RoundAndSwap.successful — Method

successful(t_stat::TerminationStatusCode)

Determine if a model was succesful

RoundAndSwap.successful_swaps — Method

successful_swaps(swapper::Swapper)

Get a list of swaps which succeeded

RoundAndSwap.swap — Method

swap(models::Array{Model}, swapper::Swapper)

Arguments:

models: An array of models, one for each thread

swapper: An already initialised swapper, this can either be clean or it can be partially complete

RoundAndSwap.swap — Method

swap(model::Model, consider_swapping::Array{VariableRef}; optimizer = nothing, max_swaps = Inf,save_path::Union{Nothing, String}=nothing)

Given a model and a list of variables swap the integer values to improve the objective function

Arguments:

models: An array of models, one for each thread

consider_swapping: An array of variables to consider swapping

optimizer: A specific optimizer to use, if the desired is not in [Gurobi, Ipopt, HiGHS]

max_swaps: The maximum number of swaps, default is Inf

save_path: A path to save the swapper to after each swap, isnothing(save_path) ? don't save : save, default is nothing

RoundAndSwap.swap — Method

swap(models::Vector{Model}, consider_swapping::Vector{VariableRef}; max_swaps = Inf, optimizer = nothing)

Given a model and a list of variables swap the integer values to improve the objective function

Arguments:

models: An Vector of models, one for each thread

consider_swapping: An Vector of variables to consider swapping

max_swaps: The maximum number of swaps, default is Inf

save_path: A path to save the swapper to after each swap, isnothing(save_path) ? don't save : save, default is nothing

auto_cpu_limit: Whether to set a cpu time limie. Once enough feasible and infeasible solutions have been determined if there is sufficient gap in the solve time between the two this will be used to set the cpu time limit.

RoundAndSwap.total_optimisation_time — Method

total_optimisation_time(swapper::Swapper)

Get the total time spent in optimizers

RoundAndSwap.try_swapping! — Method

try_swapping!(models::Array{Model}, swapper::Swapper)

Given a model and the swapper, try all swaps in swapper.to_swap

RoundAndSwap.unfix! — Method

unfix!(models::Array{Model}, swapper::Swapper)

Unfix all variables in consider swapping

RoundAndSwap.unfix! — Method

unfix!(variable::VariableRef)

unfix! a variable are reset 0 and 1 as lower and upper bounds

RoundAndSwap.unsuccessful_swaps — Method

unsuccessful_swaps(swapper::Swapper)

Get a list of swaps which failed